+++ examples/lexicons/codelingo/ast/typescript/src/common/KeyManager.ts
