+++ examples/lexicons/codelingo/ast/typescript/src/common/model.ts
